pyproject.toml
[tool.poetry]
name = "multimetro-inteligente"
version = "0.1.0"
description = "Sistema de Mapeamento e Comparação de Placas Eletrônicas"
authors = ["Poturlak <poturlak@icloud.com>"]

[tool.poetry.dependencies]
python = "^3.11"
PyQt6 = "^6.6.0"
Pillow = "^10.0.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.3"
pytest-qt = "^4.2.0"
black = "^23.10.1"
ruff = "^0.1.6"
mypy = "^1.7.0"

[tool.black]
line-length = 88

[tool.ruff]
line-length = 88
select = ["E", "W", "F", "I", "B"]
